requirements.txt
kafka-python
pyspark
requests
feedparser
beautifulsoup4
fastf1
pandas
scikit-learn
sqlalchemy
psycopg2
python-dotenv
tweepy
azure-storage-blob
emoji
langdetect
googletrans==4.0.0-rc1
google-generativeai
pymongo